@@ -13,10 +13,10 @@
     "prefer-stable": true,
     "require": {
         "php": "^7.4",
-        "laravel/framework": "^6.0|^7.0"
+        "laravel/framework": "^8.0"
     },
     "require-dev": {
-        "orchestra/testbench": "^4.0|^5.0",
+        "orchestra/testbench": "^6.0",
         "phpunit/phpunit": "^9.0"
     },
     "autoload": {
@@ -26,7 +26,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Elegant\\DataTables\\Tests\\": "tests/"
+            "Elegant\\DataTables\\Tests\\": "tests/",
+            "Elegant\\DataTables\\Tests\\Database\\Factories\\": "tests/database/factories/"
         }
     },
     "extra": {
